+++ src/Config/ConfigContainerInterface.php
@@ -68,6 +68,8 @@ interface ConfigContainerInterface
 
     /**
      * Return a list of types which are zip-able
+     *
+     * @return list<string>
      */
     public function getTypesAllowedForZip(): array;