+++ src/Module/Application/Playback/PlayAction.php
@@ -947,7 +947,7 @@ final class PlayAction implements ApplicationActionInterface
         $end          = 0;
         $range_values = sscanf(Core::get_server('HTTP_RANGE'), "bytes=%d-%d", $start, $end);
 
-        if ($range_values > 0 && ($start > 0 || $end > 0)) {
+        if (!$transcode && $range_values > 0 && ($start > 0 || $end > 0)) {
             // Calculate stream size from byte range
             if ($range_values >= 2) {
                 $end = (int)min($end, $streamConfiguration['file_size'] - 1);
@@ -961,7 +961,7 @@ final class PlayAction implements ApplicationActionInterface
                     'Content-Range header received, which we cannot fulfill due to unknown final length (transcoding?)',
                     [LegacyLogger::CONTEXT_TYPE => __CLASS__]
                 );
-            } elseif (!$transcode) {
+            } else {
                 $this->logger->debug(
                     'Content-Range header received, skipping ' . $start . ' bytes out of ' . $streamConfiguration['file_size'],
                     [LegacyLogger::CONTEXT_TYPE => __CLASS__]
@@ -1019,7 +1019,7 @@ final class PlayAction implements ApplicationActionInterface
             }
         }
 
-        if ($random || $demo_id) {
+        if ($transcode || $random || $demo_id) {
             header('Accept-Ranges: none');
         } else {
             header('Accept-Ranges: bytes');
@@ -1069,7 +1069,7 @@ final class PlayAction implements ApplicationActionInterface
         $buf_all = '';
         $r_arr   = array($filepointer);
         $w_arr   = $e_arr = array();
-        $status  = stream_select($r_arr, $w_arr, $e_arr, 2);
+        $status  = stream_select($r_arr, $w_arr, $e_arr, null);
         if ($status === false) {
             $this->logger->error(
                 'stream_select failed.',