+++ src/Module/Application/Register/ShowAddUserAction.php
@@ -30,8 +30,9 @@ use Ampache\Config\ConfigurationKeyEnum;
 use Ampache\Module\Application\ApplicationActionInterface;
 use Ampache\Module\Application\Exception\AccessDeniedException;
 use Ampache\Module\Authorization\GuiGatekeeperInterface;
+use Ampache\Module\User\Registration\RegistrationAgreementRendererInterface;
 use Ampache\Module\Util\Mailer;
-use Ampache\Module\Util\Ui;
+use Ampache\Module\Util\UiInterface;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
 
@@ -41,26 +42,40 @@ final class ShowAddUserAction implements ApplicationActionInterface
 
     private ConfigContainerInterface $configContainer;
 
+    private RegistrationAgreementRendererInterface $registrationAgreementRenderer;
+
+    private UiInterface $ui;
+
     public function __construct(
-        ConfigContainerInterface $configContainer
+        ConfigContainerInterface $configContainer,
+        RegistrationAgreementRendererInterface $registrationAgreementRenderer,
+        UiInterface $ui
     ) {
-        $this->configContainer = $configContainer;
+        $this->configContainer               = $configContainer;
+        $this->registrationAgreementRenderer = $registrationAgreementRenderer;
+        $this->ui                            = $ui;
     }
 
     /**
      * @param ServerRequestInterface $request
      * @param GuiGatekeeperInterface $gatekeeper
      * @return ResponseInterface|null
-     * @todo drop copy/paste code from register action after fixing the captcha problem
      */
     public function run(ServerRequestInterface $request, GuiGatekeeperInterface $gatekeeper): ?ResponseInterface
     {
-        /* Check Perms */
+        // Check allow_public_registration
         if (
-            $this->configContainer->isFeatureEnabled(ConfigurationKeyEnum::ALLOW_PUBLIC_REGISTRATION) === false &&
+            $this->configContainer->isFeatureEnabled(ConfigurationKeyEnum::ALLOW_PUBLIC_REGISTRATION) === false
+        ) {
+            throw new AccessDeniedException('Error `allow_public_registration` disabled');
+        }
+        // Check for confirmation email requirements when mail is disabled
+        if (
+            $this->configContainer->isFeatureEnabled(ConfigurationKeyEnum::ALLOW_PUBLIC_REGISTRATION) === true &&
+            $this->configContainer->isFeatureEnabled(ConfigurationKeyEnum::USER_NO_EMAIL_CONFIRM) === false &&
             !Mailer::is_mail_enabled()
         ) {
-            throw new AccessDeniedException('Error attempted registration');
+            throw new AccessDeniedException('Error `mail_enable` failed. Enable `user_no_email_confirm` to disable mail requirements');
         }
 
         /* Don't even include it if we aren't going to use it */
@@ -76,9 +91,14 @@ final class ShowAddUserAction implements ApplicationActionInterface
                     $this->configContainer->getWebPath()
                 )
             );
-            require_once __DIR__ . '/../../Util/Captcha/init.php';
         }
-        require_once Ui::find_template('show_user_registration.inc.php');
+
+        $this->ui->show(
+            'show_user_registration.inc.php',
+            [
+                'registrationAgreementRenderer' => $this->registrationAgreementRenderer,
+            ]
+        );
 
         return null;
     }