+++ src/Repository/Model/Search.php
@@ -114,6 +114,9 @@ class Search extends playlist_object
                     $this->$key = $value;
                 }
             }
+            if (!is_array($this->rules)) {
+                $this->rules = array();
+            }
             // make sure saved rules match the correct names
             $rule_count = 0;
             foreach ($this->rules as $rule) {